+++ include/oneapi/dpl/experimental/kernel_templates
@@ -12,6 +12,6 @@
 #include "../internal/common_config.h"
 #include "../pstl/onedpl_config.h"
 
-#include "../pstl/hetero/dpcpp/kt_esimd_radix_sort.h"
+#include "kt/esimd_radix_sort.h"
 
 #endif // _ONEDPL_KERNEL_TEMPLATES
