+++ include/oneapi/dpl/experimental/kt/esimd_radix_sort_cooperative.h
@@ -17,10 +17,10 @@
 #define _ONEDPL_kt_esimd_radix_sort_cooperative_H
 
 #include <ext/intel/esimd.hpp>
-#include "../sycl_defs.h"
-#include "../execution_sycl_defs.h"
-#include "../parallel_backend_sycl_utils.h"
-#include "../utils_ranges_sycl.h"
+#include "../../pstl/hetero/dpcpp/sycl_defs.h"
+#include "../../pstl/hetero/dpcpp/execution_sycl_defs.h"
+#include "../../pstl/hetero/dpcpp/parallel_backend_sycl_utils.h"
+#include "../../pstl/hetero/dpcpp/utils_ranges_sycl.h"
 #include <cstdint>
 
 namespace oneapi::dpl::experimental::esimd::impl
